current 2025 commit april 03/5/2025

+++ shopitems/src/main/java/org/example/domains/Stock.java
@@ -24,7 +24,7 @@ public class Stock extends PanacheEntity {
     public String store;
 
     @Column(nullable = false)
-    public Integer quantityReceived;
+    public BigDecimal quantityReceived;
 
     @Column(nullable = false)
     public BigDecimal unitCostPrice;
